Updated source code in the library
Former-commit-id: b17fde44d52b0ada7e4584a76670642cdc47f686
Former-commit-id: a8f45b47932f6051329ab3e683f00110ddd8ea19

+++ library/Assembly/Windows/Sample_Programs/Chapter_17/popcnt.c
@@ -1,7 +1,7 @@
-long popcnt_array ( long *a, int size )
+long popcnt_array ( long long *a, int size )
 {
     int w, b;
-    long word;
+    long long word;
     long n;
 
     n = 0;
